Addes SPs and new methods to DAL project

+++ OtiumActio_DbConnection/Helpers/SeedData.cs
@@ -12,15 +12,15 @@ namespace OtiumActio.Helpers
         {
             list = new List<Activity>()
             {
-                new Activity { Category = (long)Category.Naturvanrding, 
+                new Activity { Category = (int)Category.Naturvanrding, 
                     Description = "Göteborg. Träffpunkt: Ruddalen entre", 
                     Participants = 3, 
                     Date = FormatDateTime(7) },
-                new Activity { Category = (long)Category.Idrott, 
+                new Activity { Category = (int)Category.Idrott, 
                     Description = "Borås. Träffpunkt: Central station", 
                     Participants = 8, 
                     Date = FormatDateTime(10) },
-                new Activity { Category = (long)Category.Bio, 
+                new Activity { Category = (int)Category.Bio, 
                     Description = "Göteborg. Träffpunkt: HagaBion.", 
                     Participants = 2, 
                     Date = FormatDateTime(3) },
